+++ LibCarla/source/carla/sensor/data/DVSEvent.h
@@ -14,32 +14,36 @@ namespace data {
 
 #pragma pack(push, 1)
 struct DVSEvent {
-    /** Default constructor **/
+    /// Default constructor
     DVSEvent() = default;
 
-    /** Copy Constructor **/
+    /// Copy Constructor
     DVSEvent(const DVSEvent &arg)
       :x(arg.x), y(arg.y), t(arg.t), pol(arg.pol){}
 
-    /** Moving constructor **/
+    /// Moving constructor
     DVSEvent(const DVSEvent &&arg)
       :x(std::move(arg.x)), y(std::move(arg.y)), t(std::move(arg.t)), pol(std::move(arg.pol)){}
 
-    /** Constructor **/
+    /// Constructor
     DVSEvent(std::uint16_t x, std::uint16_t y, std::int64_t t, bool pol)
       : x(x), y(y), t(t), pol(pol) {}
 
-    /** Assignement operator **/
-    DVSEvent &operator=(const DVSEvent &other)
-    {
-      x = other.x; y = other.y; t = other.t; pol = other.pol;
+    /// Assignement operator
+    DVSEvent &operator=(const DVSEvent &other) {
+      x = other.x;
+      y = other.y;
+      t = other.t;
+      pol = other.pol;
       return *this;
     }
 
-    /** Move Assignement operator **/
-    DVSEvent &operator=(const DVSEvent &&other)
-    {
-      x = std::move(other.x); y = std::move(other.y); t = std::move(other.t); pol = std::move(other.pol);
+    /// Move Assignement operator
+    DVSEvent &operator=(const DVSEvent &&other) {
+      x = std::move(other.x);
+      y = std::move(other.y);
+      t = std::move(other.t);
+      pol = std::move(other.pol);
       return *this;
     }